requirements.txt
# Core AI/ML Libraries
anthropic>=0.20.0
openai>=1.0.0
transformers>=4.35.0
accelerate>=0.20.0  # Optional, for better GPU handling
tiktoken>=0.5.0
torch>=2.0.0
google-generativeai>=0.8.0

# Data Processing
numpy>=1.22.0
pandas>=2.0.0
scikit-learn>=1.2.0

# Configuration and Validation
pydantic>=2.0.0

# Logging and Utilities
python-dotenv>=1.0.0

# Development and Testing
pytest>=7.0.0
mypy>=1.0.0
black>=23.0.0
flake8>=6.0.0

# Optional: Performance Monitoring
py-spy>=0.3.14

# Optional: Visualization (if needed for complexity analysis)
matplotlib>=3.7.0
seaborn>=0.12.0